# book_networking_api/src/main/java/com/example/book_networking_api/controller/auth/AuthenticationResponse.java
package com.example.book_networking_api.controller.auth;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {
    private String token;
}