# book_networking_api/src/main/java/com/example/book_networking_api/dto/FeedbackRequestDto.java
package com.example.book_networking_api.dto;

import jakarta.validation.constraints.*;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;


@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
public class FeedbackRequestDto {

    @NotNull(message = "204" )
    private Integer bookId;
    @Positive(message = "200")
    @Min(value = 0, message = "201")
    @Max(value = 5, message = "202")
    private Double note;
    @NotNull(message = "203")
    @NotEmpty(message = "203")
    @NotBlank(message = "203")
    private String comment;
}